Modified Nest simulator for MUSIC compatibility
* When launching MUSIC applications, you need to define the number
of mpi_processes must be used by MUSIC and Nest, separately.
Therefore, here I added a part in which the number of mpi_processes
which can be used by Nest is extracted from the json.
Now, in the json file, a new line of code must specify
the number of mpi_processes the bsb can use.

* Since the number of mpi_processes bsb can use is different from
the total number of mpi_processes (at least 1 process must be used by MUSIC),
we cannot use the Barrier and/or Bcast methods of MPI library. Those methods
wait all the mpi_processes currently in use, even the process
controlled by MUSIC.